LibWeb+Compositor: Move OpenGLContext in Compositor
